LayoutTests/imported/w3c/web-platform-tests/mathml/relations/css-styling/displaystyle-2-expected.txt


PASS mtable element (lowercase) 
PASS mfrac element (lowercase) 
PASS mroot element (lowercase) 
PASS msub element (lowercase) 
PASS msup element (lowercase) 
PASS msubsup element (lowercase) 
PASS munder element (lowercase) 
PASS mover element (lowercase) 
PASS munderover element (lowercase) 
FAIL mtable element (uppercase) assert_approx_equals: cell with displaystyle true expected 50 +/- 5 but got 11
FAIL mfrac element (uppercase) assert_approx_equals: numerator expected 50 +/- 5 but got 11
FAIL mroot element (uppercase) assert_approx_equals: index expected 50 +/- 5 but got 11
FAIL msub element (uppercase) assert_approx_equals: subscript expected 50 +/- 5 but got 11
FAIL msup element (uppercase) assert_approx_equals: superscript expected 50 +/- 5 but got 11
FAIL msubsup element (uppercase) assert_approx_equals: subscript expected 50 +/- 5 but got 11
FAIL munder element (uppercase) assert_approx_equals: underscript expected 50 +/- 5 but got 11
FAIL mover element (uppercase) assert_approx_equals: overscript expected 50 +/- 5 but got 11
FAIL munderover element (uppercase) assert_approx_equals: underscript expected 50 +/- 5 but got 11
FAIL mtable element (alternate_case) assert_approx_equals: cell with displaystyle true expected 50 +/- 5 but got 11
FAIL mfrac element (alternate_case) assert_approx_equals: numerator expected 50 +/- 5 but got 11
FAIL mroot element (alternate_case) assert_approx_equals: index expected 50 +/- 5 but got 11
FAIL msub element (alternate_case) assert_approx_equals: subscript expected 50 +/- 5 but got 11
FAIL msup element (alternate_case) assert_approx_equals: superscript expected 50 +/- 5 but got 11
FAIL msubsup element (alternate_case) assert_approx_equals: subscript expected 50 +/- 5 but got 11
FAIL munder element (alternate_case) assert_approx_equals: underscript expected 50 +/- 5 but got 11
FAIL mover element (alternate_case) assert_approx_equals: overscript expected 50 +/- 5 but got 11
FAIL munderover element (alternate_case) assert_approx_equals: underscript expected 50 +/- 5 but got 11
⫿
⫿
  
⫿
⫿
 
⫿
⫿
 
⫿
⫿
 
⫿
⫿
 
⫿
⫿
⫿
 
⫿
⫿
⫿
⫿
⫿
 
⫿
⫿
 
⫿
⫿
 
⫿
⫿
⫿
